Fixes Issue 136.
Actually prooftool can now marc cut-ancestors, this issue indeed needs more, like marking subformulas, but since we never use it in our project and it reduces performance of prooftool, I'm closing this issue.

Also I did some changes in XMLExporter. There is still problem in the exporter in the sense that exported files can't be compiled by our parser because of lack of permutation rules, so we need some "clever" parser.

git-svn-id: https://gapt.googlecode.com/svn/trunk@909 85d21ef8-add0-11de-ae77-5bfb9c5a3f34