The parsing and unfolding of a resolution schemata is ready. Only the parsing of the second-order substitution is not ready, but when it is hard-coded everything works fine. A resolution refutation which verifies the correctness of the resolution schemata (see the .rs file) is also produced.
IMPORTANT: For the resolution refutation which uses a cut-rule I had to change in calculus/lk 
the usual equiality "==" between HOLFormulas with the syntaxEquals. Otherwise the 
CutRule(...) can not be applied. This happens for first time to me! 
But indeed it is correct in this case. So, syntaxEquality should be 
used everywhere in my oppinion.







git-svn-id: https://gapt.googlecode.com/svn/trunk@1307 85d21ef8-add0-11de-ae77-5bfb9c5a3f34